app/requirements.txt
flask==2.0.1
pyjwt==2.1.0